Refactor the run-on-every-ref code into something we can :sparkles: test :sparkles:
